Move member vocabulary to vocabulary module.
